+++ spring-cloud-starter-tencent-polaris-discovery/src/test/java/com/tencent/cloud/polaris/DiscoveryPropertiesAutoConfigurationTest.java
@@ -51,21 +51,6 @@ public class DiscoveryPropertiesAutoConfigurationTest {
 		});
 	}
 
-	@Test
-	public void testInit() {
-		ApplicationContextRunner applicationContextRunner = new ApplicationContextRunner().withConfiguration(
-						AutoConfigurations.of(
-								PolarisContextAutoConfiguration.class,
-								TestConfiguration.class,
-								DiscoveryPropertiesAutoConfiguration.class))
-				.withPropertyValues("spring.cloud.polaris.discovery.register=false")
-				.withPropertyValues("spring.cloud.consul.discovery.register=false")
-				.withPropertyValues("spring.cloud.consul.discovery.enabled=false");
-		applicationContextRunner.run(context -> {
-			assertThat(context).hasSingleBean(DiscoveryPropertiesAutoConfiguration.class);
-		});
-	}
-
 	@Configuration
 	static class TestConfiguration {
 		@Bean